+++ src/main/java/kuit/project/beering/controller/AuthController.java
@@ -10,8 +10,8 @@ import kuit.project.beering.service.MemberService;
 import kuit.project.beering.service.TokenService;
 import kuit.project.beering.security.jwt.JwtInfo;
 import kuit.project.beering.util.BaseResponse;
-import kuit.project.beering.util.exception.AgreementValidationException;
-import kuit.project.beering.util.exception.FieldValidationException;
+import kuit.project.beering.util.exception.validation.AgreementValidationException;
+import kuit.project.beering.util.exception.validation.FieldValidationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
