+++ src/main/java/kuit/project/beering/controller/OAuthController.java
@@ -14,8 +14,8 @@ import kuit.project.beering.security.jwt.OAuthTokenInfo;
 import kuit.project.beering.service.OAuthService;
 import kuit.project.beering.util.BaseResponse;
 import kuit.project.beering.util.BaseResponseStatus;
-import kuit.project.beering.util.exception.AgreementValidationException;
-import kuit.project.beering.util.exception.FieldValidationException;
+import kuit.project.beering.util.exception.validation.AgreementValidationException;
+import kuit.project.beering.util.exception.validation.FieldValidationException;
 import kuit.project.beering.util.exception.SignupNotCompletedException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
