+++ src/main/java/kuit/project/beering/controller/advice/AuthControllerAdvice.java
@@ -5,7 +5,7 @@ import kuit.project.beering.controller.AuthController;
 import kuit.project.beering.util.BaseResponse;
 import kuit.project.beering.util.BaseResponseStatus;
 import kuit.project.beering.dto.common.ObjectValidationError;
-import kuit.project.beering.util.exception.AgreementValidationException;
+import kuit.project.beering.util.exception.validation.AgreementValidationException;
 import kuit.project.beering.util.exception.CustomJwtException;
 import kuit.project.beering.util.exception.DuplicateNicknameException;
 import kuit.project.beering.util.exception.DuplicateUsernameException;
